% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collected-df.R
\name{get_collections}
\alias{get_collections}
\title{The collections of a collected data frame}
\usage{
get_collections(x, call = rlang::caller_env())
}
\arguments{
\item{x}{A \link{collected_df}.}

\item{call}{The calling environment, used to report errors. Defaults to the
caller of \code{get_collections()}.}
}
\value{
A list of \link{collection} objects.
}
\description{
Returns the list of \link{collection}s carried by a \link{collected_df} (including a
\link{conjoint_df}, which is a \code{collected_df}). Each collection names a group of
columns that belong together - for example, the conjoint attributes of a
\code{conjoint_df}.
}
\examples{
# A conjoint_df carries one collection per attribute.
cjt <- conjoint_df(example_utilities, example_crosswalk)
get_collections(cjt)
}
