% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{is_ordered}
\alias{is_ordered}
\title{Is a collection ordered?}
\usage{
is_ordered(x)
}
\arguments{
\item{x}{A \link{collection} object.}
}
\value{
A single logical.
}
\description{
Is a collection ordered?
}
